Refactor httpserver integration tests into per-domain files
The OAuth integration tests were all in a single 4,389-line file
(oauth_integration_test.go) containing one OAuthFlowSuite with 88 test
methods spanning many unrelated domains. Split it into focused files
that share a common base:

  - integration_common_test.go  (suite setup, shared helpers, fixtures)
  - oauth_flow_test.go
  - oidc_test.go
  - mfa_test.go
  - email_verification_flow_test.go
  - password_reset_flow_test.go
  - authorize_consent_test.go
  - introspection_revocation_test.go
  - client_credentials_test.go
  - admin_test.go

All 88 test methods stay on OAuthFlowSuite, so suite discovery is
unchanged.

Fixes inconsistent build-tag gating:
  - oauth_integration_test.go had its //go:build integration tag
    commented out ("// /* //go:build integration */"), so the
    container-backed suite ran on every `go test`, while
    avatar_integration_test.go (real tag) never ran in CI.
  - All integration files now carry a live //go:build integration tag.
  - avatar_integration_test.go referenced an undefined
    generateCodeChallenge and so had never compiled under the tag; the
    helper is now defined once in integration_common_test.go and reused
    by both suites (replacing the inline S256 computation).

Wires integration tests into automation:
  - Makefile: add `test-integration` target (go test -tags integration).
  - CI: add a dedicated "Run integration tests" step that runs the
    tagged suites (testcontainers needs the runner's Docker daemon).

Minor cleanup: email_verification_test.go replaces hand-rolled
contains/findSubstring helpers with strings.Contains.
