+++ .github/workflows/pull_request.yaml
@@ -34,8 +34,23 @@ jobs:
             -----END EC PRIVATE KEY-----
           JWT_ISSUER: http://localhost:8080
 
-      - name: Run tests
+      - name: Run unit tests
         run: go test -v ./...
+        env:
+          # Don't worry, this is a key used only for testing.
+          JWT_PRIVATE_KEY: |
+            -----BEGIN EC PRIVATE KEY-----
+            MHcCAQEEICQMNHONu2Sud2tu6jgOZs3LIj5yOZr89NBMLYiyqBK/oAoGCCqGSM49
+            AwEHoUQDQgAERCHWHrX20emk31HypGNgptwBjdZOyBybV/9BLTbJPj8UsZ/46ri5
+            /eFKkRfNApxFU/5lk1RGQJqt8t0GvkkJdw==
+            -----END EC PRIVATE KEY-----
+          JWT_ISSUER: http://localhost:8080
+
+      # Integration tests are gated behind the `integration` build tag. They use
+      # testcontainers to spin up Postgres/MinIO, which relies on the Docker
+      # daemon available on GitHub-hosted ubuntu runners.
+      - name: Run integration tests
+        run: go test -v -tags integration -timeout 600s ./...
         env:
           # Don't worry, this is a key used only for testing.
           JWT_PRIVATE_KEY: |
