+++ Makefile
@@ -48,6 +48,11 @@ migrate-down:
 test:
 	go test ./...
 
+# Integration tests are gated behind the `integration` build tag because they
+# spin up Postgres/MinIO containers via testcontainers and require Docker.
+test-integration:
+	go test -tags integration ./...
+
 lint:
 	go vet ./...
